Rename BayesianRecommender file from base.py to core.py
